requirements.txt
imutils==0.5.4
numpy==1.19.5
picamera==1.13
Pillow==9.5.0
pytesseract==0.3.10